フォルダごとのコマンド分割法に変更し既存コマンドを移行
