+++ SerialController/Commands/McuCommands/InfinityWatt.py
@@ -1,0 +1,11 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+
+from ..McuCommandBase import McuCommand
+
+# Mash A button
+class InfinityWatt(McuCommand):
+	NAME = '無限ワット'
+
+	def __init__(self, sync_name = 'inf_watt'):
+		super(InfinityWatt, self).__init__(sync_name)
