+++ SerialController/Commands/PythonCommands/AutoLeague.py
@@ -1,19 +1,21 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 
-from .. import PythonCommandBase as base
+from ..PythonCommandBase import PythonCommand, ImageProcPythonCommand
+from Commands.Keys import KeyPress, Button, Direction, Stick
 
 # Auto league
 # 自動リーグ周回(画像認識なし)
-class AutoLeague(base.PythonCommand):
-	def __init__(self, name):
-		super(AutoLeague, self).__init__(name)
+class AutoLeague(PythonCommand):
+	NAME = '自動リーグ周回'
+
+	def __init__(self):
+		super(AutoLeague, self).__init__()
 
 	def do(self):
 		self.hold(Direction(Stick.LEFT, 70))
-		while True:
-			self.wait(0.5)
 
+		while True:
 			for _ in range(0, 10):
 				self.press(Button.A, wait=0.5)
